BigIntegerFiel Telephone

+++ clients/models.py
@@ -5,5 +5,5 @@ from django.db import models
 class Client(models.Model):
     name = models.CharField(max_length=200)
     address = models.CharField(max_length=200)
-    telephone = models.IntegerField()
+    telephone = models.BigIntegerField()
     create_time=models.DateTimeField(auto_now_add=True)
